refactor: change MAX_BACKUPS type to size_t and update related code

+++ src/constants.h
@@ -7,7 +7,7 @@
 #define MAX_STRING_SIZE 40
 #define MAX_JOB_FILE_NAME_SIZE 256
 
-extern unsigned int MAX_BACKUPS; // argv[2]
-extern size_t MAX_THREADS;       // argv[3]
+extern size_t MAX_BACKUPS; // argv[2]
+extern size_t MAX_THREADS; // argv[3]
 
 #endif // CONSTANTS_H
